Run the game locally: setup runbook, test launchers, VFS and account fixes
Documents and scripts the path from a clean checkout to a running local
game (PM-011), which the PM confirmed working on 2026-09-24.

- doc/ai/topics/project/local-setup-runbook.md: verified procedure and
  every trap hit on the way. The project needs the v142 toolset
  (MSVC 14.29) and must build inside a 14.29 developer environment,
  otherwise the cc crate compiles SQLite with the newest MSVC and the Rust
  link in common's cargo pre-build fails. The runbook also covers the
  Windows SDK override, running cargo from src/, game data from the
  README's MEGA archive (a packed client, unpacked into data/ for the
  servers), the database migrations and the server config.
- run/: double-click launchers to start/stop the servers and the client
  and to create an account. They call Windows tools by full path and keep
  CRLF endings (run/.gitattributes).
- src/tools/vfs-browser: roselib reads VFS offsets as i32, and widening
  with `as u64` sign-extended every offset past 2 GB, so extracting the
  2.36 GB rose.vfs failed at file 32,300 of 36,264. Reinterpret the fields
  as u32 like triggervfs does.
- scripts/create-account.ps1: enforce the client's login limits (ID 6-30,
  password 8-30). It accepted 1-character names, creating accounts the
  client refuses with "Email too short".
- doc/ai records: decisions PM-011..PM-018 and IMP-012..IMP-019, state,
  handoff and session record.

+++ src/tools/vfs-browser/src/vfs.rs
@@ -154,9 +154,12 @@ impl Vfs {
                         out.push(VfsEntry {
                             vfs_index: i,
                             path: f.filepath.to_string_lossy().replace('\\', "/"),
-                            offset: f.offset as u64,
-                            size: f.size as u64,
-                            block_size: f.block_size as u64,
+                            // roselib reads these as i32; the on-disk fields are
+                            // unsigned (triggervfs reads them that way), so a plain
+                            // `as u64` sign-extends every offset past 2 GB.
+                            offset: f.offset as u32 as u64,
+                            size: f.size as u32 as u64,
+                            block_size: f.block_size as u32 as u64,
                             is_deleted: f.is_deleted,
                             is_compressed: f.is_compressed,
                             is_encrypted: f.is_encrypted,
